+++ android/src/main/kotlin/com/example/imagegallerysaver/ImageGallerySaverPlugin.kt
@@ -1,13 +1,10 @@
 package com.example.imagegallerysaver
 
-import androidx.annotation.NonNull
-import android.annotation.TargetApi
 import android.content.ContentValues
 import android.content.Context
 import android.content.Intent
 import android.graphics.Bitmap
 import android.graphics.BitmapFactory
-import android.media.MediaScannerConnection
 import android.net.Uri
 import android.os.Environment
 import android.os.Build
@@ -18,220 +15,145 @@ import io.flutter.plugin.common.MethodCall
 import io.flutter.plugin.common.MethodChannel
 import io.flutter.plugin.common.MethodChannel.MethodCallHandler
 import io.flutter.plugin.common.MethodChannel.Result
-import io.flutter.plugin.common.PluginRegistry.Registrar
+//import io.flutter.plugin.common.PluginRegistry.Registrar
 import java.io.File
 import java.io.FileInputStream
 import java.io.IOException
 import android.text.TextUtils
 import android.webkit.MimeTypeMap
-import java.io.OutputStream
+
 
 class ImageGallerySaverPlugin : FlutterPlugin, MethodCallHandler {
-    private lateinit var methodChannel: MethodChannel
     private var applicationContext: Context? = null
+    private var methodChannel: MethodChannel? = null
 
-    override fun onAttachedToEngine(@NonNull binding: FlutterPlugin.FlutterPluginBinding) {
-        this.applicationContext = binding.applicationContext
-        methodChannel = MethodChannel(binding.binaryMessenger, "image_gallery_saver")
-        methodChannel.setMethodCallHandler(this)
-    }
 
-    override fun onMethodCall(@NonNull call: MethodCall,@NonNull result: Result): Unit {
-        when (call.method) {
-            "saveImageToGallery" -> {
-                val image = call.argument<ByteArray?>("imageBytes")
-                val quality = call.argument<Int?>("quality")
-                val name = call.argument<String?>("name")
+//    companion object {
+//        @JvmStatic
+//        fun registerWith(registrar: Registrar) {
+//            val instance = ImageGallerySaverPlugin()
+//            instance.onAttachedToEngine(registrar.context(), registrar.messenger())
+//        }
+//    }
 
-                result.success(
-                    saveImageToGallery(
-                        BitmapFactory.decodeByteArray(
-                            image ?: ByteArray(0),
-                            0,
-                            image?.size ?: 0
-                        ), quality, name
-                    )
-                )
+    override fun onMethodCall(call: MethodCall, result: Result): Unit {
+        when {
+            call.method == "saveImageToGallery" -> {
+                val image = call.argument<ByteArray>("imageBytes") ?: return
+                val quality = call.argument<Int>("quality") ?: return
+                val name = call.argument<String>("name")
+
+                result.success(saveImageToGallery(BitmapFactory.decodeByteArray(image, 0, image.size), quality, name))
             }
-
-            "saveFileToGallery" -> {
-                val path = call.argument<String?>("file")
-                val name = call.argument<String?>("name")
+            call.method == "saveFileToGallery" -> {
+                val path = call.argument<String>("file") ?: return
+                val name = call.argument<String>("name")
                 result.success(saveFileToGallery(path, name))
             }
-
             else -> result.notImplemented()
         }
+
     }
 
-    override fun onDetachedFromEngine(@NonNull binding: FlutterPlugin.FlutterPluginBinding) {
-        applicationContext = null
-        methodChannel.setMethodCallHandler(null);
-    }
 
-    private fun generateUri(extension: String = "", name: String? = null): Uri? {
+    private fun generateUri(extension: String = "", name: String? = null): Uri {
         var fileName = name ?: System.currentTimeMillis().toString()
-        val mimeType = getMIMEType(extension)
-        val isVideo = mimeType?.startsWith("video")==true
+        
+        if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.Q) {
+            var uri = MediaStore.Images.Media.EXTERNAL_CONTENT_URI
 
-        return if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.Q) {
-            // >= android 10
-            val uri = when {
-                isVideo -> MediaStore.Video.Media.EXTERNAL_CONTENT_URI
-                else -> MediaStore.Images.Media.EXTERNAL_CONTENT_URI
-            }
-
-            val values = ContentValues().apply {
-                put(MediaStore.MediaColumns.DISPLAY_NAME, fileName)
-                put(
-                    MediaStore.MediaColumns.RELATIVE_PATH, when {
-                        isVideo -> Environment.DIRECTORY_MOVIES
-                        else -> Environment.DIRECTORY_PICTURES
-                    }
-                )
-                if (!TextUtils.isEmpty(mimeType)) {
-                    put(when {isVideo -> MediaStore.Video.Media.MIME_TYPE
-                        else -> MediaStore.Images.Media.MIME_TYPE
-                    }, mimeType)
+            val values = ContentValues()
+            values.put(MediaStore.MediaColumns.DISPLAY_NAME, fileName)
+            values.put(MediaStore.MediaColumns.RELATIVE_PATH, Environment.DIRECTORY_PICTURES)
+            val mimeType = getMIMEType(extension)
+            if (!TextUtils.isEmpty(mimeType)) {
+                values.put(MediaStore.Images.Media.MIME_TYPE, mimeType)
+                if (mimeType!!.startsWith("video")) {
+                    uri = MediaStore.Video.Media.EXTERNAL_CONTENT_URI
+                    values.put(MediaStore.MediaColumns.RELATIVE_PATH, Environment.DIRECTORY_MOVIES)
                 }
             }
-
-            applicationContext?.contentResolver?.insert(uri, values)
-
+            return applicationContext?.contentResolver?.insert(uri, values)!!
         } else {
-            // < android 10
-            val storePath =
-                Environment.getExternalStoragePublicDirectory(when {
-                    isVideo -> Environment.DIRECTORY_MOVIES
-                    else -> Environment.DIRECTORY_PICTURES
-                }).absolutePath
-            val appDir = File(storePath).apply {
-                if (!exists()) {
-                    mkdir()
-                }
+            val storePath = Environment.getExternalStorageDirectory().absolutePath + File.separator + Environment.DIRECTORY_PICTURES
+            val appDir = File(storePath)
+            if (!appDir.exists()) {
+                appDir.mkdir()
             }
-
-            val file =
-                File(appDir, if (extension.isNotEmpty()) "$fileName.$extension" else fileName)
-            Uri.fromFile(file)
+            if (extension.isNotEmpty()) {
+                fileName += (".$extension")
+            }
+            return Uri.fromFile(File(appDir, fileName))
         }
     }
 
-    /**
-     * get file Mime Type
-     *
-     * @param extension extension
-     * @return file Mime Type
-     */
     private fun getMIMEType(extension: String): String? {
-        return if (!TextUtils.isEmpty(extension)) {
-            MimeTypeMap.getSingleton().getMimeTypeFromExtension(extension.lowercase())
-        } else {
-            null
+        var type: String? = null;
+        if (!TextUtils.isEmpty(extension)) {
+            type = MimeTypeMap.getSingleton().getMimeTypeFromExtension(extension.toLowerCase())
         }
+        return type
     }
 
-    /**
-     * Send storage success notification
-     *
-     * @param context context
-     * @param fileUri file path
-     */
-    private fun sendBroadcast(context: Context, fileUri: Uri?) {
-        if (Build.VERSION.SDK_INT < Build.VERSION_CODES.Q) {
-            val mediaScanIntent = Intent(Intent.ACTION_MEDIA_SCANNER_SCAN_FILE)
-            mediaScanIntent.data = fileUri
-            context.sendBroadcast(mediaScanIntent)
-        }
-    }
-
-    private fun saveImageToGallery(
-        bmp: Bitmap?,
-        quality: Int?,
-        name: String?
-    ): HashMap<String, Any?> {
-        // check parameters
-        if (bmp == null || quality == null) {
-            return SaveResultModel(false, null, "parameters error").toHashMap()
-        }
-        // check applicationContext
+    private fun saveImageToGallery(bmp: Bitmap, quality: Int, name: String?): HashMap<String, Any?> {
         val context = applicationContext
-            ?: return SaveResultModel(false, null, "applicationContext null").toHashMap()
-        var fileUri: Uri? = null
-        var fos: OutputStream? = null
-        var success = false
-        try {
-            fileUri = generateUri("jpg", name = name)
-            if (fileUri != null) {
-                fos = context.contentResolver.openOutputStream(fileUri)
-                if (fos != null) {
-                    println("ImageGallerySaverPlugin $quality")
-                    bmp.compress(Bitmap.CompressFormat.JPEG, quality, fos)
-                    fos.flush()
-                    success = true
-                }
-            }
-        } catch (e: IOException) {
-            SaveResultModel(false, null, e.toString()).toHashMap()
-        } finally {
-            fos?.close()
+        val fileUri = generateUri("jpg", name = name)
+        return try {
+            val fos = context?.contentResolver?.openOutputStream(fileUri)!!
+            println("ImageGallerySaverPlugin $quality")
+            bmp.compress(Bitmap.CompressFormat.JPEG, quality, fos)
+            fos.flush()
+            fos.close()
+            context!!.sendBroadcast(Intent(Intent.ACTION_MEDIA_SCANNER_SCAN_FILE, fileUri))
             bmp.recycle()
-        }
-        return if (success) {
-            sendBroadcast(context, fileUri)
             SaveResultModel(fileUri.toString().isNotEmpty(), fileUri.toString(), null).toHashMap()
-        } else {
-            SaveResultModel(false, null, "saveImageToGallery fail").toHashMap()
-        }
-    }
-
-    private fun saveFileToGallery(filePath: String?, name: String?): HashMap<String, Any?> {
-        // check parameters
-        if (filePath == null) {
-            return SaveResultModel(false, null, "parameters error").toHashMap()
-        }
-        val context = applicationContext ?: return SaveResultModel(
-            false,
-            null,
-            "applicationContext null"
-        ).toHashMap()
-        var fileUri: Uri? = null
-        var outputStream: OutputStream? = null
-        var fileInputStream: FileInputStream? = null
-        var success = false
-
-        try {
-            val originalFile = File(filePath)
-            if(!originalFile.exists()) return SaveResultModel(false, null, "$filePath does not exist").toHashMap()
-            fileUri = generateUri(originalFile.extension, name)
-            if (fileUri != null) {
-                outputStream = context.contentResolver?.openOutputStream(fileUri)
-                if (outputStream != null) {
-                    fileInputStream = FileInputStream(originalFile)
-
-                    val buffer = ByteArray(10240)
-                    var count = 0
-                    while (fileInputStream.read(buffer).also { count = it } > 0) {
-                        outputStream.write(buffer, 0, count)
-                    }
-
-                    outputStream.flush()
-                    success = true
-                }
-            }
         } catch (e: IOException) {
             SaveResultModel(false, null, e.toString()).toHashMap()
-        } finally {
-            outputStream?.close()
-            fileInputStream?.close()
-        }
-        return if (success) {
-            sendBroadcast(context, fileUri)
-            SaveResultModel(fileUri.toString().isNotEmpty(), fileUri.toString(), null).toHashMap()
-        } else {
-            SaveResultModel(false, null, "saveFileToGallery fail").toHashMap()
         }
     }
+
+    private fun saveFileToGallery(filePath: String, name: String?): HashMap<String, Any?> {
+        val context = applicationContext
+        return try {
+            val originalFile = File(filePath)
+            val fileUri = generateUri(originalFile.extension, name)
+
+            val outputStream = context?.contentResolver?.openOutputStream(fileUri)!!
+            val fileInputStream = FileInputStream(originalFile)
+
+            val buffer = ByteArray(10240)
+            var count = 0
+            while (fileInputStream.read(buffer).also { count = it } > 0) {
+                outputStream.write(buffer, 0, count)
+            }
+
+            outputStream.flush()
+            outputStream.close()
+            fileInputStream.close()
+
+            context!!.sendBroadcast(Intent(Intent.ACTION_MEDIA_SCANNER_SCAN_FILE, fileUri))
+            SaveResultModel(fileUri.toString().isNotEmpty(), fileUri.toString(), null).toHashMap()
+        } catch (e: IOException) {
+            SaveResultModel(false, null, e.toString()).toHashMap()
+        }
+    }
+
+    override fun onAttachedToEngine(binding: FlutterPlugin.FlutterPluginBinding) {
+        onAttachedToEngine(binding.applicationContext, binding.binaryMessenger)
+    }
+
+    override fun onDetachedFromEngine(binding: FlutterPlugin.FlutterPluginBinding) {
+        applicationContext = null
+        methodChannel!!.setMethodCallHandler(null);
+        methodChannel = null;
+    }
+
+    private fun onAttachedToEngine(applicationContext: Context, messenger: BinaryMessenger) {
+        this.applicationContext = applicationContext
+        methodChannel = MethodChannel(messenger, "image_gallery_saver")
+        methodChannel!!.setMethodCallHandler(this)
+    }
+
 }
 
 class SaveResultModel(var isSuccess: Boolean,